requirements.txt
asgiref==3.4.1
# certifi==2018.8.13
chardet==3.0.4
# db-helper==2.0.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.7
psycopg2==2.9.1
psycopg2-binary==2.8.6
# pytz==2021.1
requests==2.19.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.23
uWSGI==2.0.17.1
# whitenoise==5.3.0
# pycrypto==2.6.1
celery==5.0.1
setuptools==57.4.0
